requirements.txt
decorator==4.2.1
Django==2.0.1
django-extensions==1.9.9
djangorestframework==3.7.7
gunicorn==19.7.1
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.4
wcwidth==0.1.7
